+++ Towerdefense/Security.env
@@ -1,0 +1,4 @@
+HOST=serverless-eastus.sysp0000.db3.skysql.com
+USER=dbpbf26371036
+PASS=kI8Y02vu6ZJyr4?82NZMeR
+DB=game_db
