+++ Towerdefense/login.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#ifndef LOGIN_H
+#define LOGIN_H
+
+// Declare the login function
+bool login();
+
+#endif
